# Cache the packageName to which receiver belongs