Cache the packageName to which receiver belongs in java/lang/Class.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>