@DanHeidinga 2021-11-15
@r30shah Can you also describe why this field is worth caching?  I'm assuming it's for performance reasons but it would help to identify the benchmarks where this is a benefit