@r30shah 2021-11-16
@DanHeidinga , yes, we were looking into the cases where Method.invoke [1] calls in JDK11 lands in `slowVerifyAccess` [2] due to missing single level caller/receiver cache [3] we keep on AccessibleObject. While looking into the code-path where we spend significant amount of time in such case, I found that we call the `getPackageName` on Class object and compute the name of the package everytime. In such case, cached package name will be benefitial.
@joransiu  has prepared a simple benchmark using JMH framework to simulate such scenarios, which I am uploading here. [jmh-benchmark-joran.zip](https://github.com/eclipse-openj9/openj9/files/7549215/jmh-benchmark-joran.zip)
. From the set of benchmarks, microbenchmark `ReflectMethodInvokeBenchmark.testDifferentCallersInvokeStaticMethod` simulates such scenario where multiple classes calls a static method using same Method object forcing us to miss the single level caller/receiver cache on Accessible Object. Result is in the following table comparing baseline with the chage to cache the packageName.


| Benchmark | Baseline Throughput Performance | Cache PackageName Throughput Performance |
|-----------|---------------------------------|------------------------------------------|
| ReflectMethodInvokeBenchmark.testDifferentCallersInvokeStaticMethod | 100 | 415.63 |

[1]. https://github.com/ibmruntimes/openj9-openjdk-jdk11/blob/84f7b67e27b02ad1c8fe6f5d230fc28fc9025a39/src/java.base/share/classes/java/lang/reflect/Method.java#L549-L567
[2]. https://github.com/ibmruntimes/openj9-openjdk-jdk11/blob/84f7b67e27b02ad1c8fe6f5d230fc28fc9025a39/src/java.base/share/classes/java/lang/reflect/AccessibleObject.java#L630
[3]. https://github.com/ibmruntimes/openj9-openjdk-jdk11/blob/84f7b67e27b02ad1c8fe6f5d230fc28fc9025a39/src/java.base/share/classes/java/lang/reflect/AccessibleObject.java#L611-L616