@DanHeidinga 2021-11-16
@r30shah thanks for the JMH benchmark.  Is there any indication this is hitting real world workloads or other benchmarks?  There's a footprint / throughput tradeoff happening here and it would be good to know how prevalent this pattern is in the wild.

As an additional change, it probably makes sense to `intern()` the packageName String so there is only 1 copy in memory rather than 1 per class from the package.