@joransiu 2021-11-16
@DanHeidinga : This performance opportunity arose from our analysis of a customer reported issue - highlighted as a Java 11 performance difference vs Java 8 in their application.   The main cause of difference is due to Java 11's removal of `Reflection.quickCheckMemberAccess(clazz, modifiers)` from Method.invoke(), which had guarded calls to `getCallerClass()` and `checkAccess()`.  Now, under certain conditions, we end up stressing these previously guarded paths more often.

For this customer app, the main calls to Method.invoke that hits the slow verify path are from their application code.   However, the next hottest callers of Method.invoke are from from: `java.security.Provider.newInstanceUtil` driven by various call stacks related to Jar verifier and Log4J code.   The caching of package name will help with those call stacks too.