@r30shah 2021-11-17
> As an additional change, it probably makes sense to intern() the packageName String so there is only 1 copy in memory rather than 1 per class from the package.

@DanHeidinga  I was looking into it while creating change, and found that the packageName string was created and `intern()` in [1] when we compute the non-ArrayClass package name which is what we call it the class is not primitive.  

[1]. https://github.com/eclipse-openj9/openj9/blob/69ff1a02db3b0ca40fa8ed2db2cbfdc6be29afb4/jcl/src/java.base/share/classes/java/lang/Class.java#L2116-L2123