@tajila 2021-11-19
Jenkins test sanity zlinux jdk8,jdk17